MatrixMorphism_abstract.{charpoly,det,fcp,trace}: Move to category
